created database; created model; created repository interfaces; created dto; created mapper; created part of service; created part of controller;
WORKING REST + log4j;

running app without error;

+++ src/main/java/ro/msg/learning/shop/exception/StockNotFoundException.java
@@ -1,7 +1,7 @@
 package ro.msg.learning.shop.exception;
 
 public class StockNotFoundException extends RuntimeException {
-    public StockNotFoundException(Long id) {
+    public StockNotFoundException(Integer id) {
         super("Could not found supplier with id: " + id);
     }
 }
